+++ Engine/Core/Source/Memory/Allocator.cpp
@@ -4,19 +4,19 @@
 
 namespace Luma
 {
-    void* FDefaultAllocator::allocate(size_t size, size_t alignment)
+    void* FDefaultAllocator::allocate(uint64_t size, uint64_t alignment)
     {
         return ::operator new(size);
     }
 
-    void* FDefaultAllocator::callocate(size_t size, size_t alignment)
+    void* FDefaultAllocator::callocate(uint64_t size, uint64_t alignment)
     {
         void* data = ::operator new(size);
         Memory::memset(data, 0, size);
         return data;
     }
 
-    void* FDefaultAllocator::reallocate(void* ptr, size_t size, size_t alignment)
+    void* FDefaultAllocator::reallocate(void* ptr, uint64_t size, uint64_t alignment)
     {
         return ::operator new (size, ptr);
     }
